add tests with enabled feature
